Fixed return type of _normalize not working with cv2

+++ hep2_classification/preprocessing.py
@@ -60,7 +60,7 @@ def _denoise(img: np.ndarray, h: float = DENOISE_STRENGTH) -> np.ndarray:
 
 def _normalize(img: np.ndarray, result_range: Tuple[int, int] = (0, 255)) -> np.ndarray:
     """ Normalizes given image by performing simple interpolation to given range. """
-    return np.interp(img, (img.min(), img.max()), result_range).astype(np.uint)
+    return np.interp(img, (img.min(), img.max()), result_range).astype(np.uint8)
 
 
 def _equalize_histogram(img: np.ndarray) -> np.ndarray:
